+ add mypy + add bootstrap.bat * refactor TelegramParts
